+++ lucene/core/src/java/org/apache/lucene/index/DocumentsWriterPerThread.java
@@ -616,7 +616,7 @@ final class DocumentsWriterPerThread implements Accountable {
   @Override
   public long ramBytesUsed() {
     assert lock.isHeldByCurrentThread();
-    return (deleteDocIDs.length * Integer.BYTES)
+    return (deleteDocIDs.length * (long) Integer.BYTES)
         + pendingUpdates.ramBytesUsed()
         + indexingChain.ramBytesUsed();
   }
